+++ src/pooled_connection.rs
@@ -2,7 +2,7 @@ use std::cell::RefCell;
 use std::collections::HashMap;
 use std::time::{Duration, Instant};
 
-use dbus::{BusName, ConnPath, Connection, Message, Path};
+use dbus::{BusName, ConnPath, Connection, Member, Message, Path};
 
 use extensions::DurationExtensions;
 use player::MPRIS2_PATH;
@@ -19,11 +19,16 @@ const NAME_HAS_OWNER_TIMEOUT: i32 = 100; // ms
 
 impl PooledConnection {
     pub(crate) fn new(connection: Connection) -> Self {
+        // Subscribe to events that relate to players. See `is_watched_message` below for more
+        // details.
         let _ = connection.add_match(
             "interface='org.freedesktop.DBus.Properties',member='PropertiesChanged',path='/org/mpris/MediaPlayer2'",
         );
         let _ = connection.add_match(
             "interface='org.mpris.MediaPlayer2.Player',member='Seeked',path='/org/mpris/MediaPlayer2'",
+        );
+        let _ = connection.add_match(
+            "type='signal',sender='org.freedesktop.DBus',interface='org.freedesktop.DBus',member='NameOwnerChanged'",
         );
         PooledConnection {
             connection: connection,
@@ -75,8 +80,17 @@ impl PooledConnection {
             .and_then(|reply| reply.get1())
     }
 
-    pub(crate) fn last_event_for_unique_name(&self, unique_name: &str) -> Option<Instant> {
-        self.last_event.borrow().get(unique_name).cloned()
+    /// Returns `true` is an event has been recorded for the given bus name, after the given
+    /// instant.
+    ///
+    /// If no event have been seen at all for the given bus name, or the last event was on or
+    /// before the provided instant then `false` will be returned.
+    pub(crate) fn is_bus_updated_after(&self, bus_name: &str, after: &Instant) -> bool {
+        self.last_event
+            .borrow()
+            .get(bus_name)
+            .map(|updated_at| updated_at > after)
+            .unwrap_or(false)
     }
 
     pub(crate) fn process_events_blocking(&self, duration: Duration) {
@@ -93,42 +107,90 @@ impl PooledConnection {
             if ms_left < 2 {
                 break;
             }
-            match self.connection.incoming(ms_left as u32).next() {
-                Some(message) => {
-                    if PooledConnection::is_watched_message(&message) {
-                        self.process_message(&message);
-                    }
-                }
-                None => {
-                    // Time is up. No more messages.
-                    break;
+            if let Some(message) = self.connection
+                .incoming(ms_left as u32)
+                .filter(PooledConnection::is_watched_message)
+                .next()
+            {
+                self.process_message(&message);
+            }
+        }
+    }
+
+    /// Block until a MPRIS2 event for the given unique bus name is detected, or the bus disappears
+    /// (the program exits).
+    ///
+    /// Events for other buses will also be recorded, but the method will not return until a
+    /// matching one has been found or the bus disappears.
+    pub(crate) fn process_events_blocking_until_dirty(&self, unique_name: &str) {
+        // mpris2 library must have a timeout, but since this function calls it in a loop it
+        // doesn't really matter what limit we set.
+        const LOOP_INTERVAL_MS: u32 = 1000;
+        let start = Instant::now();
+
+        loop {
+            for message in self.connection
+                .incoming(LOOP_INTERVAL_MS)
+                .filter(PooledConnection::is_watched_message)
+            {
+                self.process_message(&message);
+                if self.is_bus_updated_after(unique_name, &start) {
+                    return;
                 }
             }
         }
     }
 
+    /// Returns true if a given D-Bus Message is a signal that is interesting for this library.
+    ///   1. MPRIS2 "Seeked" signal.
+    ///   2. D-Bus "PropertiesChanged" signal on a MPRIS2 path.
+    ///   3. D-Bus "NameOwnerChanged" signal, to detect when players disappear.
+    ///
+    /// Since the connection given to the PooledConnection could already have subscriptions that
+    /// are not listed in this file, it is important that messages are manually filtered before
+    /// acting on them.
     fn is_watched_message(message: &Message) -> bool {
-        use std::ops::Deref;
-
-        if let Some(message_path) = message.path().as_ref() {
-            if message_path.deref() != MPRIS2_PATH {
-                return false;
-            }
+        if message.sender() == Some(BusName::from("org.freedesktop.DBus")) {
+            message.member() == Some(Member::from("NameOwnerChanged"))
+        } else if message.path() == Some(Path::from(MPRIS2_PATH)) {
+            message.member() == Some(Member::from("PropertiesChanged"))
+                || message.member() == Some(Member::from("Seeked"))
+        } else {
+            false
         }
-
-        if let Some(message_member) = message.member().as_ref() {
-            if message_member.deref() == "Seeked" || message_member.deref() == "PropertiesChanged" {
-                return true;
-            }
-        }
-
-        false
     }
 
+    /// Takes a message and updates the latest update time of the sender bus.
     fn process_message(&self, message: &Message) {
-        message
-            .sender()
-            .map(|unique_name| self.mark_bus_as_updated((*unique_name).to_owned()));
+        if message.member() == Some(Member::from("NameOwnerChanged")) {
+            let _ = self.process_name_owner_changed_event(message);
+        } else {
+            // Process mpris signal
+            message
+                .sender()
+                .map(|unique_name| self.mark_bus_as_updated((*unique_name).to_owned()));
+        }
+    }
+
+    fn process_name_owner_changed_event(
+        &self,
+        message: &Message,
+    ) -> Result<(), ::dbus::arg::TypeMismatchError> {
+        let mut iter = message.iter_init();
+        let name: String = iter.read()?;
+
+        if name.starts_with("org.mpris.") {
+            let old_name: String = iter.read()?;
+            let new_name: String = iter.read()?;
+
+            // If "new_name" is empty, then the bus disappeared. "Wake" any potentially waiting
+            // loop up.
+            if new_name.is_empty() {
+                self.mark_bus_as_updated(old_name);
+            }
+        }
+
+        Ok(())
     }
 
     fn mark_bus_as_updated<S: Into<String>>(&self, bus_name: S) {